Crate nd_jpegli_sys

Structs

J_COLOR_SPACE

J_DCT_METHOD

jpegli_component_info

jpegli_compress_struct

jpegli_decompress_struct

jpegli_error_mgr

jpegli_memory_manager

jpegli_progress_mgr

jpegli_source_mgr

Constants

FALSE

JPEGLI_HEADER_OK

JPEGLI_HEADER_TABLES_ONLY

JPEGLI_SUSPENDED

JPEG_EOI

TRUE

Functions

nd_jpegli_create_compress⚠

Initialize a new compress context.

nd_jpegli_create_decompress⚠

Initialize a new decompress context.

nd_jpegli_destroy_compress⚠

Destroy a compress context.

nd_jpegli_destroy_decompress⚠

Destroy a decompress context.

nd_jpegli_finish_compress⚠

Finish compressing.

nd_jpegli_finish_decompress⚠

Finish decompression.

nd_jpegli_free_err_str⚠

Free an error string allocated by this library.

nd_jpegli_mem_dest⚠

Setup a compress context to use the a buffer as a file destination.

nd_jpegli_mem_src⚠

Setup a decompress context to use the a buffer as a file source.

nd_jpegli_read_header⚠

Read a jpeg’s headers, and set compression parameters.

nd_jpegli_read_scanlines⚠

Read scanlines from a decompress context.

nd_jpegli_set_defaults⚠

Set default parameters for a compress operation.

nd_jpegli_set_quality⚠

Set the compression quality.

nd_jpegli_set_xyb_mode⚠

Set xyb mode.

nd_jpegli_start_compress⚠

Start compressing.

nd_jpegli_start_decompress⚠

Start decompressing.

nd_jpegli_write_scanlines⚠

Write scanlines to a compress context.

Type Aliases

JDIMENSION

JSAMPARRAY

boolean

c_char

Equivalent to C’s char type.

c_int

Equivalent to C’s signed int (int) type.

c_uint

Equivalent to C’s unsigned int type.

c_ulong

Equivalent to C’s unsigned long type.

j_compress_ptr

j_decompress_ptr

Unions

jpegli_error_mgr_msg_parm